Rename the package, extend the unit test placeholders, write comments.
Renamed the current unit test package to include the 'test' directory.
This is useful in JavaDoc so that the coding and testing packages are
listed uniquely in the 'Packages' section of the auto-generate HTML page

Coded placeholders for test functions.

Wrote comments for the TestGameManage class and one of the unit tests.